Move `FormCollectionPicker` to collections module
